optimise bandwidth usage
This change was made to optimise bandwidth usage by ensuring that the
emoji image files are only sent once per session and usage. Prior to
this commit every response from the Phoenix LiveView WebSocket would
include all used emojis in each payload. Now I optimised it by keeping
track of all emojis that have been sent already and setting phx-update
to only append new emojis.

I am not 100% satisfied with the solution yet, because I ended up with
inline style elements in an empty div container. It resulted in this
because phx-update=append requires ids to know what is there and what
is new. So this is something I might look into modifying in the future
but will require some more research.